refactor TrustAllCertificates as separate class add snellen_chart.pdf test file, including its license comment in README
